requirements.txt
numpy==1.24.3
pandas==1.5.3
python-docx==0.8.11
flask==2.3.3 